relocate subprocess_popen_text + subprocess_terminate to easybuild.tools.run
