+++ easybuild/tools/py2vs3/py3.py
@@ -59,6 +59,8 @@ try:
 except ImportError:
     HAVE_DISTUTILS = False
 
+from easybuild.tools.run import subprocess_popen_text, subprocess_terminate
+
 # string type that can be used in 'isinstance' calls
 string_type = str
 
@@ -73,24 +75,6 @@ def json_loads(body):
         body = body.decode('utf-8', 'ignore')
 
     return json.loads(body)
-
-
-def subprocess_popen_text(cmd, **kwargs):
-    """Call subprocess.Popen in text mode with specified named arguments."""
-    # open stdout/stderr in text mode in Popen when using Python 3
-    kwargs.setdefault('stderr', subprocess.PIPE)
-    return subprocess.Popen(cmd, stdout=subprocess.PIPE, universal_newlines=True, **kwargs)
-
-
-def subprocess_terminate(proc, timeout):
-    """Terminate the subprocess if it hasn't finished after the given timeout"""
-    try:
-        proc.communicate(timeout=timeout)
-    except subprocess.TimeoutExpired:
-        for pipe in (proc.stdout, proc.stderr, proc.stdin):
-            if pipe:
-                pipe.close()
-        proc.terminate()
 
 
 def raise_with_traceback(exception_class, message, traceback):
